+++ src-tauri/src/lib.rs
@@ -406,10 +406,10 @@ pub fn run() {
             audio::commands::audio_set_crossfade,
             audio::commands::audio_set_gapless,
             audio::commands::audio_set_normalization,
-            audio::commands::audio_list_devices,
-            audio::commands::audio_canonicalize_selected_device,
-            audio::commands::audio_default_output_device_name,
-            audio::commands::audio_set_device,
+            audio::device_commands::audio_list_devices,
+            audio::device_commands::audio_canonicalize_selected_device,
+            audio::device_commands::audio_default_output_device_name,
+            audio::device_commands::audio_set_device,
             audio::commands::audio_chain_preload,
             discord::discord_update_presence,
             discord::discord_clear_presence,
